fx: move RaisingBlock
