+++ src/global/vars.c
@@ -9,7 +9,7 @@
 #include "game/effect_routines/flood.h"
 #include "game/effect_routines/lara_effects.h"
 #include "game/effect_routines/powerup.h"
-#include "game/objects/effects/raising_block.h"
+#include "game/effect_routines/raising_block.h"
 #include "game/objects/effects/sand.h"
 #include "game/objects/effects/stairs2slope.h"
 #include "game/objects/effects/turn_180.h"
@@ -19,11 +19,11 @@ char *GameMemoryPointer = NULL;
 int32_t g_FPSCounter = 0;
 
 void (*g_EffectRoutines[])(ITEM_INFO *item) = {
-    Turn180,          FX_DinoStomp,   FX_LaraNormal,
-    FX_Bubbles,       FX_FinishLevel, FX_Earthquake,
-    FX_Flood,         RaisingBlock,   Stairs2Slope,
-    DropSand,         FX_PowerUp,     FX_Explosion,
-    FX_LaraHandsFree, FX_FlipMap,     FX_LaraDrawRightGun,
+    Turn180,          FX_DinoStomp,    FX_LaraNormal,
+    FX_Bubbles,       FX_FinishLevel,  FX_Earthquake,
+    FX_Flood,         FX_RaisingBlock, Stairs2Slope,
+    DropSand,         FX_PowerUp,      FX_Explosion,
+    FX_LaraHandsFree, FX_FlipMap,      FX_LaraDrawRightGun,
     FX_ChainBlock,    FX_Flicker,
 };
